[Gold IV] Title: 부분합, Time: 8 ms, Memory: 2300 KB -BaekjoonHub

+++ 백준/Gold/1806. 부분합/부분합.cpp
@@ -8,7 +8,7 @@ int main() {
   int n, s;
   cin >> n >> s;
 
-  vector<int> v(n);
+  vector<int16_t> v(n);
 
   for (int i = 0; i < n; ++i) {
     cin >> v[i];
